removed ViewController and Endpoint testing changed to tests suite
