+++ Solidus-Swift-SDK/MainViewController.swift
@@ -1,0 +1,22 @@
+//
+//  MainViewController.swift
+//  Solidus-Swift-SDK
+//
+//  Created by Omar Guzman on 5/3/16.
+//  Copyright © 2016 Omar Guzman. All rights reserved.
+//
+
+import UIKit
+
+class MainViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+    
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+}
