+++ Solidus-Swift-SDKTests/Solidus_Swift_SDKTests.swift
@@ -21,15 +21,47 @@ class Solidus_Swift_SDKTests: XCTestCase {
         super.tearDown()
     }
     
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    func testCountries()
+    {
+        let asyncExpectation = expectationWithDescription("async call")
+        
+        Countries.List { (result) in
+            if((result as? Bool) != nil)
+            {
+                XCTAssertFalse(result as! Bool, "service error")
+                asyncExpectation.fulfill()
+            }
+            else
+            {
+                XCTAssertNotNil(result, "service passed")
+                //asyncExpectation.fulfill()
+            }
+            
+        }
+        self.waitForExpectationsWithTimeout(10) { (error) in
+            XCTAssertNil(error, "something went wrong")
+        }
     }
     
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measureBlock {
-            // Put the code you want to measure the time of here.
+    func testCountryShow()
+    {
+        let asyncExpectation = expectationWithDescription("async call")
+        
+        Countries.Show("117") { (result) in
+            if((result as? Bool) != nil)
+            {
+                XCTAssertFalse(result as! Bool, "service error")
+                asyncExpectation.fulfill()
+            }
+            else
+            {
+                XCTAssertNotNil(result, "service passed")
+                asyncExpectation.fulfill()
+            }
+            
+        }
+        self.waitForExpectationsWithTimeout(10) { (error) in
+            XCTAssertNil(error, "something went wrong")
         }
     }
